Small fixes.

+++ libsrc/pylith/topology/FieldOps.cc
@@ -211,7 +211,7 @@ pylith::topology::FieldOps::getSubfieldNamesDomain(const pylith::topology::Field
         err = PetscDSGetDiscretization(fieldDS, iField, &discretization);PYLITH_CHECK_ERROR(err);
         PylithInt fieldIndex = -1;
         err = PetscDSGetFieldIndex(fieldDS, discretization, &fieldIndex);PYLITH_CHECK_ERROR(err);
-        assert(fieldIndex >= 0 && fieldIndex < subfieldNames.size());
+        assert(fieldIndex >= 0 && size_t(fieldIndex) < subfieldNames.size());
         subfieldNamesDomain[iField] = subfieldNames[fieldIndex];
     } // for
